ruzstd
Module blocks

In a Zstandard frame, there’s a frame header, followed by one or more blocks.

A block contains data, and a header describing how that data is encoded, as well as other misc metadata.

https://github.com/facebook/zstd/blob/dev/doc/zstd_compression_format.md#blocks

Modules

block

Block header definitions.

literals_section

Utilities and representations for the first half of a block, the literals section. It contains data that is then copied from by the sequences section.

sequence_section

Utilities and representations for the second half of a block, the sequence section. This section copies literals from the literals section into the decompressed output.