ruzstd::blocks
Module block

Block header definitions.

Structs

BlockHeader

A representation of a single block header. As well as containing a frame header, each Zstandard frame contains one or more blocks.

Enums

BlockType

There are 4 different kinds of blocks, and the type of block influences the meaning of Block_Size.