ruzstd
Module decoding

Structures and utilities used for reading from data, decoding that data and storing the output.

Modules

bit_reader

bit_reader_reverse

block_decoder

decodebuffer

dictionary

literals_section_decoder

This module contains the [decompress_literals] function, used to take a parsed literals header and a source and decompress it.

scratch

Structures that wrap around various decoders to make decoding easier.

sequence_execution

sequence_section_decoder