ruzstd::decoding::decodebuffer
Struct DecodeBuffer

pub struct DecodeBuffer {
    pub dict_content: Vec<u8>,
    pub window_size: usize,
    pub hash: XxHash64,
    /* private fields */
}

Fields
dict_content: Vec<u8>window_size: usizehash: XxHash64
Implementations

impl DecodeBuffer

pub fn new(window_size: usize) -> DecodeBuffer

pub fn reset(&mut self, window_size: usize)

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn push(&mut self, data: &[u8])

pub fn repeat( &mut self, offset: usize, match_length: usize, ) -> Result<(), DecodeBufferError>

pub fn can_drain_to_window_size(&self) -> Option<usize>

Check if and how many bytes can currently be drawn from the buffer

pub fn can_drain(&self) -> usize

pub fn drain_to_window_size(&mut self) -> Option<Vec<u8>>

Drain as much as possible while retaining enough so that decoding si still possible with the required window_size At best call only if can_drain_to_window_size reports a โhighโ number of bytes to reduce allocations

pub fn drain_to_window_size_writer( &mut self, sink: impl Write, ) -> Result<usize, Error>

pub fn drain(&mut self) -> Vec<u8>

drain the buffer completely

pub fn drain_to_writer(&mut self, sink: impl Write) -> Result<usize, Error>

pub fn read_all(&mut self, target: &mut [u8]) -> Result<usize, Error>