ruzstd::decoding
Module dictionary

Structs

Dictionary

Zstandard includes support for “raw content” dictionaries, that store bytes optionally used during sequence execution.

Enums

DictionaryDecodeError

Constants

MAGIC_NUM

This 4 byte (little endian) magic number refers to the start of a dictionary