Module ruzstd::decoding::sequence_section_decoder

Enums

DecodeSequenceError

Constants

LL_MAX_LOG

“The maximum allowed accuracy log for literals length and match length tables is 9”

ML_MAX_LOG

“The maximum allowed accuracy log for literals length and match length tables is 9”

OF_MAX_LOG

“The maximum accuracy log for the offset table is 8.”

Functions

decode_sequences

Decode the provided source as a series of sequences into the supplied target.