ruzstd

Module encoding

Structures and utilities used for compressing/encoding data into the Zstd format.

Structs

FrameCompressor

An interface for compressing arbitrary data with the ZStandard compression algorithm.

MatchGeneratorDriver

This is the default implementation of the Matcher trait. It allocates and reuses the buffers when possible.

Enums

CompressionLevel

The compression mode used impacts the speed of compression, and resulting compression ratios. Faster compression will result in worse compression ratios, and vice versa.

Sequence

Sequences that a Matcher can produce

Traits

Matcher

Trait used by the encoder that users can use to extend the matching facilities with their own algorithm making their own tradeoffs between runtime, memory usage and compression ratio

Functions

compress

Convenience function to compress some source into a target without reusing any resources of the compressor

compress_to_vec

Convenience function to compress some source into a Vec without reusing any resources of the compressor