ruzstd
Module frame_decoder

Framedecoder is the man struct users interact with to decode zstd frames

Zstandard compressed data is made of one or more [Frame]s. Each frame is independent and can be decompressed independently of other frames. This module contains structures and utilities that can be used to decode a frame.

Structs

FrameDecoder

This implements a decoder for zstd frames.

Enums

BlockDecodingStrategy

FrameDecoderError