Module ruzstd::fse

FSE, short for Finite State Entropy, is an encoding technique that assigns shorter codes to symbols that appear more frequently in data, and longer codes to less frequent symbols.

FSE works by mutating a state and using that state to index into a table.

Zstandard uses two different kinds of entropy encoding: FSE, and Huffman coding. Huffman is used to compress literals, while FSE is used for all other symbols (literal length code, match length code, offset code).

https://github.com/facebook/zstd/blob/dev/doc/zstd_compression_format.md#fse

https://arxiv.org/pdf/1311.2540

Structs

Entry

A single entry in an FSE table.

FSEDecoder

FSETable

FSE decoding involves a decoding table that describes the probabilities of all literals from 0 to the highest present one

Enums

FSEDecoderError

FSETableError