ruzstd

Module io

Re-exports of std values for when the std is available.

Structs

Error

The error type for I/O operations of the Read, Write, Seek, and associated traits.

Enums

ErrorKind

A list specifying general categories of I/O error.

Traits

Read

The Read trait allows for reading bytes from a source.

Write

A trait for objects which are byte-oriented sinks.