Struct ruzstd::streaming_decoder::StreamingDecoder

pub struct StreamingDecoder<READ: Read, DEC: BorrowMut<FrameDecoder>> {
    pub decoder: DEC,
    /* private fields */
}

High level Zstandard frame decoder that can be used to decompress a given Zstandard frame.

This decoder implements io::Read, so you can interact with it by calling io::Read::read_to_end / io::Read::read_exact or passing this to another library / module as a source for the decoded content

If you need more control over how decompression takes place, you can use the lower level FrameDecoder, which allows for greater control over how decompression takes place but the implementor must call FrameDecoder::decode_blocks repeatedly to decode the entire frame.

Caveat

StreamingDecoder expects the underlying stream to only contain a single frame, yet the specification states that a single archive may contain multiple frames.

To decode all the frames in a finite stream, the calling code needs to recreate the instance of the decoder and handle crate::frame::ReadFrameHeaderError::SkipFrame errors by skipping forward the length amount of bytes, see https://github.com/KillingSpark/zstd-rs/issues/57

// `read_to_end` is not implemented by the no_std implementation.
#[cfg(feature = "std")]
{
    use std::fs::File;
    use std::io::Read;
    use ruzstd::{StreamingDecoder};

    // Read a Zstandard archive from the filesystem then decompress it into a vec.
    let mut f: File = todo!("Read a .zstd archive from somewhere");
    let mut decoder = StreamingDecoder::new(f).unwrap();
    let mut result = Vec::new();
    Read::read_to_end(&mut decoder, &mut result).unwrap();
}

Fields
decoder: DEC
Implementations

impl<READ: Read, DEC: BorrowMut<FrameDecoder>> StreamingDecoder<READ, DEC>

pub fn new_with_decoder( source: READ, decoder: DEC ) -> Result<StreamingDecoder<READ, DEC>, FrameDecoderError>

impl<READ: Read> StreamingDecoder<READ, FrameDecoder>

pub fn new( source: READ ) -> Result<StreamingDecoder<READ, FrameDecoder>, FrameDecoderError>

impl<READ: Read, DEC: BorrowMut<FrameDecoder>> StreamingDecoder<READ, DEC>

pub fn get_ref(&self) -> &READ

Gets a reference to the underlying reader.

pub fn get_mut(&mut self) -> &mut READ

Gets a mutable reference to the underlying reader.

It is inadvisable to directly read from the underlying reader.

pub fn into_inner(self) -> READ
where READ: Sized,

Destructures this object into the inner reader.

pub fn into_parts(self) -> (READ, DEC)
where READ: Sized,

Destructures this object into both the inner reader and FrameDecoder.

pub fn into_frame_decoder(self) -> DEC

Destructures this object into the inner FrameDecoder.