Crate ruzstd

Re-exports

pub use frame_decoder::BlockDecodingStrategy;

pub use frame_decoder::FrameDecoder;

pub use streaming_decoder::StreamingDecoder;

Modules

blocks

In a Zstandard frame, there’s a frame header, followed by one or more blocks.

decoding

Structures and utilities used for reading from data, decoding that data and storing the output.

frame

frame_decoder

Zstandard compressed data is made of one or more [Frame]s. Each frame is independent and can be decompressed independently of other frames. This module contains structures and utilities that can be used to decode a frame.

fse

FSE, short for Finite State Entropy, is an encoding technique that assigns shorter codes to symbols that appear more frequently in data, and longer codes to less frequent symbols.

huff0

io

Re-exports of std values for when the std is available.

streaming_decoder

Constants

VERBOSE