ruzstd
Module frame

Structs

Frame

Zstandard compressed data is made of one or more Frames. Each frame is independent and can be decompressed independently of other frames.

FrameDescriptor

The first byte is called the Frame Header Descriptor, and it describes what other fields are present.

FrameHeader

A frame header has a variable size, with a minimum of 2 bytes, and a maximum of 14 bytes.

Enums

FrameDescriptorError

FrameHeaderError

ReadFrameHeaderError

Constants

MAGIC_NUM

This magic number is included at the start of a single Zstandard frame

MAX_WINDOW_SIZE

The maximum window size is 3.75TB

MIN_WINDOW_SIZE

The minimum window size is defined as 1 KB

Functions

read_frame_header

Read a single serialized frame from the reader and return a tuple containing the parsed frame and the number of bytes read.